+++ domain/src/main/java/ru/dpopkov/knownics/domain/TextType.java
@@ -1,0 +1,9 @@
+package ru.dpopkov.knownics.domain;
+
+/**
+ * Type of translation text.
+ */
+public enum TextType {
+    PLAINTEXT,
+    MARKDOWN
+}