+++ domain/src/main/java/ru/dpopkov/knownics/domain/Translation.java
@@ -1,0 +1,10 @@
+package ru.dpopkov.knownics.domain;
+
+public interface Translation {
+
+    Language getLanguage();
+
+    TextType getType();
+
+    String getText();
+}