+++ domain/src/main/java/ru/dpopkov/knownics/domain/answer/AnswerType.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.knownics.domain.answer;
+
+public enum AnswerType {
+    ORIGINAL,
+    COMPOSED
+}