+++ domain/src/main/java/ru/dpopkov/knownics/domain/answer/Source.java
@@ -1,0 +1,52 @@
+package ru.dpopkov.knownics.domain.answer;
+
+import ru.dpopkov.knownics.domain.ModifiableEntity;
+
+public class Source extends ModifiableEntity {
+
+    private String title;
+    private String fullTitle;
+    private SourceType type;
+    private String url;
+    private String description;
+
+    public String getTitle() {
+        return title;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public String getFullTitle() {
+        return fullTitle;
+    }
+
+    public void setFullTitle(String fullTitle) {
+        this.fullTitle = fullTitle;
+    }
+
+    public SourceType getType() {
+        return type;
+    }
+
+    public void setType(SourceType type) {
+        this.type = type;
+    }
+
+    public String getUrl() {
+        return url;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+}