+++ domain/src/main/java/ru/dpopkov/knownics/domain/answer/SourceType.java
@@ -1,0 +1,10 @@
+package ru.dpopkov.knownics.domain.answer;
+
+public enum SourceType {
+    API_DOC,
+    BOOK,
+    TUTORIAL,
+    VIDEO_COURSE,
+    PERSON,
+    WEB_SITE
+}