+++ history.md
@@ -1,6 +1,8 @@
 History of implementation
 -------------------------
 
+##### 2021-12-07
+* Implement POJO data model. Closes #2.
 ##### 2021-12-05
 * Create Multi-Module Project for Domain module. Closes #3.
 ##### 2021-12-02